% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_database.R
\name{load_database}
\alias{load_database}
\title{Load trait database into R console}
\usage{
load_database(
  doi = NULL,
  version = NULL,
  path = "data/SGtraits",
  update = FALSE
)
}
\arguments{
\item{doi}{character string - doi of particular version}

\item{version}{character string - version number of database}

\item{path}{file path to where the database will be downloaded}

\item{update}{if TRUE, database versions .json will be re-downloaded}
}
\value{
a large list containing data tables
}
\description{
Currently this is a slightly modification of the \code{load_austraits}
function tailored for SGtraits.
}
\examples{
\dontrun{
SGtraits <- load_database(version = "0.0.2", path = "data/SGtraits")
}
}
\seealso{
get_versions get_version_latest
}
